Address review comments from trace refactoring PRs
