+++ homeassistant/components/trace/websocket_api.py
@@ -23,12 +23,12 @@ from homeassistant.helpers.script import (
     debug_stop,
 )
 
-from .trace import DATA_TRACE, get_all_debug_traces, get_debug_trace, get_debug_traces
+from .const import DATA_TRACE
 from .utils import TraceJSONEncoder
 
 # mypy: allow-untyped-calls, allow-untyped-defs
 
-TRACE_DOMAINS = ["automation", "script"]
+TRACE_DOMAINS = ("automation", "script")
 
 
 @callback
@@ -57,14 +57,24 @@ def async_setup(hass: HomeAssistant) -> None:
     }
 )
 def websocket_trace_get(hass, connection, msg):
-    """Get an automation or script trace."""
+    """Get an script or automation trace."""
     key = (msg["domain"], msg["item_id"])
     run_id = msg["run_id"]
 
-    trace = get_debug_trace(hass, key, run_id)
+    trace = hass.data[DATA_TRACE][key][run_id]
     message = websocket_api.messages.result_message(msg["id"], trace)
 
     connection.send_message(json.dumps(message, cls=TraceJSONEncoder, allow_nan=False))
+
+
+def get_debug_traces(hass, key):
+    """Return a serializable list of debug traces for an script or automation."""
+    traces = []
+
+    for trace in hass.data[DATA_TRACE].get(key, {}).values():
+        traces.append(trace.as_short_dict())
+
+    return traces
 
 
 @callback
@@ -72,18 +82,22 @@ def websocket_trace_get(hass, connection, msg):
 @websocket_api.websocket_command(
     {
         vol.Required("type"): "trace/list",
-        vol.Inclusive("domain", "id"): vol.In(TRACE_DOMAINS),
-        vol.Inclusive("item_id", "id"): str,
+        vol.Required("domain", "id"): vol.In(TRACE_DOMAINS),
+        vol.Optional("item_id", "id"): str,
     }
 )
 def websocket_trace_list(hass, connection, msg):
-    """Summarize automation and script traces."""
-    key = (msg["domain"], msg["item_id"]) if "item_id" in msg else None
+    """Summarize script and automation traces."""
+    domain = msg["domain"]
+    key = (domain, msg["item_id"]) if "item_id" in msg else None
 
     if not key:
-        traces = get_all_debug_traces(hass, summary=True)
+        traces = []
+        for key in hass.data[DATA_TRACE]:
+            if key[0] == domain:
+                traces.extend(get_debug_traces(hass, key))
     else:
-        traces = get_debug_traces(hass, key, summary=True)
+        traces = get_debug_traces(hass, key)
 
     connection.send_result(msg["id"], traces)
 
@@ -230,7 +244,7 @@ def websocket_subscribe_breakpoint_events(hass, connection, msg):
     }
 )
 def websocket_debug_continue(hass, connection, msg):
-    """Resume execution of halted automation or script."""
+    """Resume execution of halted script or automation."""
     key = (msg["domain"], msg["item_id"])
     run_id = msg["run_id"]
 
@@ -250,7 +264,7 @@ def websocket_debug_continue(hass, connection, msg):
     }
 )
 def websocket_debug_step(hass, connection, msg):
-    """Single step a halted automation or script."""
+    """Single step a halted script or automation."""
     key = (msg["domain"], msg["item_id"])
     run_id = msg["run_id"]
 
@@ -270,7 +284,7 @@ def websocket_debug_step(hass, connection, msg):
     }
 )
 def websocket_debug_stop(hass, connection, msg):
-    """Stop a halted automation or script."""
+    """Stop a halted script or automation."""
     key = (msg["domain"], msg["item_id"])
     run_id = msg["run_id"]
